+++ strato_query/__init__.py
@@ -12,4 +12,5 @@ December 26, 2018
 from .api_query import SDAPIQuery
 from .authentication import authenticate_to_api
 from .filters import *
+from .aggregations import *
 from .query_structures import *